+++ tests/sentry/test_sentry_helper.py
@@ -1,9 +1,6 @@
-import copy
-
 from django.test import TestCase
 
-from ambient_toolbox.sentry.helpers import SentryEventScrubber, strip_sensitive_data_from_sentry_event
-from tests.sentry.mock_data import SCRUBBED_SENTRY_EVENT, SENTRY_EVENT
+from ambient_toolbox.sentry.helpers import strip_sensitive_data_from_sentry_event
 
 
 class SentryHelperTest(TestCase):
@@ -20,16 +17,17 @@ class SentryHelperTest(TestCase):
     def test_strip_sensitive_data_from_sentry_event_missing_key_ip_address(self):
         event = {"user": {"email": "mymail@example.com", "username": "my-user"}}
 
-        self.assertIsInstance(strip_sensitive_data_from_sentry_event(event, None), dict)
+        stripped_event = strip_sensitive_data_from_sentry_event(event, None)
+
+        self.assertIsInstance(stripped_event, dict)
+        self.assertNotIn("email", stripped_event["user"].keys())
+        self.assertNotIn("username", stripped_event["user"].keys())
 
     def test_strip_sensitive_data_from_sentry_event_missing_key_username(self):
         event = {"user": {"email": "mymail@example.com", "ip_address": "127.0.0.1"}}
 
-        self.assertIsInstance(strip_sensitive_data_from_sentry_event(event, None), dict)
+        stripped_event = strip_sensitive_data_from_sentry_event(event, None)
 
-    def test_sentry_event_scrubber_on_a_regular_sentry_event(self):
-        event = copy.deepcopy(SENTRY_EVENT)
-        self.assertEqual(
-            SCRUBBED_SENTRY_EVENT,
-            SentryEventScrubber().scrub_sensitive_data_from_sentry_event(event, None),
-        )
+        self.assertIsInstance(stripped_event, dict)
+        self.assertNotIn("email", stripped_event["user"].keys())
+        self.assertNotIn("ip_address", stripped_event["user"].keys())